Country only set address on map
"Set to my entered address" does not work when only country is selected. The country field uses country codes (ex: nl,uk,de etc). The google maps api does not know how to convert those. My suggestion would be to use the text of the selected option and send that to the google maps api. This seems to work better.